+++ src/main/java/apiquality/sonar/asyncapi/checks/RulesLists.java
@@ -58,7 +58,8 @@ public final class RulesLists {
             AAR009DeclaredTagCheck.class,
             AAR010DocumentedTagCheck.class,
             AAR040DefinedChannelServersCheck.class,
-            AAR041ComponetChannelServerCheck.class           
+            AAR041ComponetChannelServerCheck.class,
+            AAR054ClassificationValidValuesCheck.class
         );
     }